Version 1 widget templates completed.
