+++ cilantro_audit/templates/audit_button.py
@@ -10,13 +10,13 @@ kivy.require(KIVY_REQUIRED_VERSION)
 Builder.load_file(os.path.dirname(os.path.abspath(__file__)) + "/widgets/audit_button.kv")
 
 
-class AuditButtonTemplate(Button):
+class AuditButton(Button):
     pass
 
 
 class RunApp(App):
     def build(self):
-        return AuditButtonTemplate()
+        return AuditButton()
 
 
 if __name__ == '__main__':
