+++ cilantro_audit/templates/audit_label.py
@@ -11,13 +11,13 @@ kivy.require(KIVY_REQUIRED_VERSION)
 Builder.load_file(os.path.dirname(os.path.abspath(__file__)) + "/widgets/audit_label.kv")
 
 
-class AuditLabelTemplate(Label):
+class AuditLabel(Label):
     pass
 
 
 class RunApp(App):
     def build(self):
-        return AuditLabelTemplate()
+        return AuditLabel()
 
 
 if __name__ == '__main__':
