+++ cilantro_audit/templates/widgets/audit_button.kv
@@ -1,13 +1,12 @@
 
-<AuditButtonTemplate@Button>:
-    # Label Attributes
+<AuditButton@Button>:
     text: 'Button Template'
     halign: 'center'
     valign: 'middle'
-    text_size:
+    text_size: self.width-20, self.height-10
     font_name:
     font_family:
-    font_size: self.height/2
+    font_size: self.height/3
     color: 0.9, 0.9, 0.9, 1.0
     disabled_color: 0, 0, 0, 0
     bold: 0
@@ -36,9 +35,4 @@
     font_features:
     base_direction:
     text_language:
-
-    # Button Attributes
     background_color: 0.5, 0.5, 0.5, 1.0
-    on_state: root.manager.transition.duration = 0.3
-    on_press: root.manager.transition.duration = 0.3
-    on_release: root.manager.transition.duration = 0.3
