+++ cilantro_audit/templates/widgets/audit_label.kv
@@ -1,5 +1,5 @@
 
-<AuditLabelTemplate@Label>:
+<AuditLabel@Label>:
     text: 'Button Template'
     halign: 'center'
     valign: 'middle'
