--- cilantro_audit/templates/widgets/page_layout.kv
@@ -1,52 +1,0 @@
-
-<AuditPageTemplate>:
-    GridLayout:
-        id: 'page'
-        rows: 3
-        cols: 1
-
-        canvas.before:
-            Color:
-                rgba: 0.0, 0.0, 0.0, 1.0
-            Rectangle:
-                pos: self.pos
-                size: self.size
-
-        GridLayout:
-            id: 'header'
-            name: 'header-grid'
-            rows: 1
-            cols: 3
-            size_hint: 1, 0.1
-
-            AuditButtonTemplate:
-                text: 'Return to Homepage'
-
-            Label:
-                text: 'Audit Page Template'
-
-            AuditButtonTemplate:
-                text: 'Exit'
-                on_release:
-                    exit(1)
-
-        GridLayout:
-            id: 'body'
-            name: 'body-grid'
-            cols: 3
-            size_hint: 1, 0.8
-
-        GridLayout:
-            id: 'footer'
-            name: 'footer-grid'
-            rows: 1
-            cols: 3
-            size_hint: 1, 0.1
-
-            AuditButtonTemplate:
-                text: 'Back'
-                on_release:
-                    root.manager.transition.direction = 'right'
-
-            AuditButtonTemplate:
-                text: 'Refresh Page'
